자잘한 주석 삭제 및 user소개 수정, 삭제시 token과 req의 유저아이디 비교 추가

+++ back/src/routers/userRouter.js
@@ -2,12 +2,6 @@ import is from "@sindresorhus/is";
 import { Router } from "express";
 import { login_required } from "../middlewares/login_required";
 import { userAuthService } from "../services/userService";
-// import { projectService } from "../services/projectService";
-// import { otherService } from "../services/otherService";
-// import { EducationService } from "../services/educationService";
-// import { certificateService } from "../services/certificateService";
-// import { AwardService } from "../services/awardService";
-// import { AboutService } from "../services/aboutService";
 
 const userAuthRouter = Router();
 
@@ -49,7 +43,7 @@ userAuthRouter.post("/login", async (req, res, next) => {
         if (user.errorMessage) {
             throw new Error(user.errorMessage);
         }
-
+        
         res.status(200).json(user);
     } catch (error) {
         next(error);
@@ -57,6 +51,7 @@ userAuthRouter.post("/login", async (req, res, next) => {
 });
 
 userAuthRouter.get("/list", login_required, async (req, res, next) => {
+    
     try {
         const users = await userAuthService.getUsers();
         res.status(200).json(users);
@@ -90,6 +85,10 @@ userAuthRouter.put("/:id", login_required, async (req, res, next) => {
         const password = req.body.password ?? null;
         const description = req.body.description ?? null;
         const image_url = req.body.image_url ?? null;
+
+        if (req.currentUserId !== user_id) {
+            throw new Error("접근권한이 없습니다.");
+        }
 
         const toUpdate = { name, email, password, description, image_url };
         const updatedUser = await userAuthService.setUser({
@@ -129,7 +128,6 @@ userAuthRouter.get(
     login_required,
     async (req, res, next) => {
         try {
-            // 전체 사용자 목록을 얻음
             const Model = req.params.type;
             const pieceword = req.params.pieceword;
 
@@ -144,13 +142,10 @@ userAuthRouter.get(
                     pieceword,
                 }); //* 이메일검색 완료
             } else if (modelName.includes(Model)) {
-                resultList = await userAuthService.getUsersWithRestrict3({ 
-                    Model, 
-                    pieceword, 
-                });
+                const errorMessage = " 추후 구현할 예정입니다.";
+                res.status.json(errorMessage);
             } else {
                 const errorMessage = "검색하려는 정확한 타입을 골라주세요.";
-                //! type에 다른 요소가 들어갈 경우 400 Bad Request가 뜨는데 임의로 string 전달.
                 res.status(400).json(errorMessage);
             }
 
@@ -158,12 +153,17 @@ userAuthRouter.get(
         } catch (error) {
             next(error);
         }
-    },
+    }
 );
 
 userAuthRouter.delete("/:id", login_required, async (req, res, next) => {
     try {
         const user_id = req.params.id;
+
+        if (req.currentUserId !== user_id) {
+            throw new Error("접근권한이 없습니다.");
+        }
+
         const deletdUser = await userAuthService.deleteUser({ user_id });
 
         if (deletdUser.errorMessage) {
